% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_wheel.R
\name{draw_wheel}
\alias{draw_wheel}
\title{Create helical wheels to visualize alpha helical sequences.}
\usage{
draw_wheel(sequence, col = c("grey", "yellow", "blue", "red"),
  labels = FALSE, label.col = "black", fixed = TRUE,
  legend = FALSE)
}
\arguments{
\item{sequence}{character vector containing amino acid sequence
from N-terminus to C-terminus}

\item{col}{colors for each amino acid type in the following order:
nonpolar residues, polar residues, basic residues, acidic residues}

\item{labels}{logical value; if TRUE, one-letter residue codes are
overlaid on the residue circles}

\item{label.col}{character value for color of labels added if `labels = TRUE`}

\item{fixed}{if TRUE (default), ensures that residues will be circles
(not ellipses) even if graphics device is rectangular}

\item{legend}{if TRUE, adds legend to plot}
}
\description{
This function visualizes alpha-helical peptides as
helical wheels.
Sequences between 2 and 18 (inclusive) characters can be visualized.
The residue in the upper half of the wheel lying at the horizontal
center representing the amino acid at the N-terminus (first in `sequence`).
}
\examples{
draw_wheel("GIGAVLKVLTTGLPALIS")
draw_wheel("QQRKRKIWSILAPLGTTL")
}
