% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleSeq}
\alias{SampleSeq}
\title{Two short amino acid sequences to visualize.}
\format{A data frame with 2 rows and 2 variables.
\describe{
  \item{Name}{name of sequence from which each sequence prefix was taken}
  \item{Sequence}{sequence prefix for visualization}
}}
\usage{
SampleSeq
}
\description{
A data set containing a name and residue sequence
for two sequences for example visualization.
}
\keyword{datasets}
