% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helixvis.R
\docType{package}
\name{helixvis}
\alias{helixvis}
\alias{helixvis-package}
\title{Quick and easy visualization of alpha-helical sequences.}
\description{
The helixvis package allows visualization of alpha-helical peptide
sequences in the form of helical wheels or Wenxiang diagrams. Although
online servers currently exist for both visualizations, all currently
require entering sequences and storing images individually. This package
allows users to create visualizations for entire databases of amino acid
sequences in the form of PNGs.
}
