% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sequence}
\alias{sequence}
\title{Amino acid sequences of 5 antimicrobial peptides}
\format{A data frame with 5 rows and 2 variables:
\describe{
  \item{Name}{antimicrobial peptide name}
  \item{Seq}{antimicrobial residue sequence}
}}
\source{
\url{http://aps.unmc.edu/AP/}
}
\usage{
sequence
}
\description{
A dataset containing the name and residue sequence of
five antimicrobial peptides from the Antimicrobial
Peptide Database (APD3).
}
\details{
See Wang et al. (2016) <doi:10.1093/nar/gkv1278> for more information.
}
\keyword{datasets}
